Modified Gear class
Method to calculate Gear Inches  introduced 
A private method setWheel to set the value of wheel diameter
Assuming values for rim and tire diameter if passed to the constructor are in inches. Also assuming a value for Cog and Chainring is always passed to the constructor.
The code should be easy to change. Next task unknown